Fix clippy warnings about type complexity in RegexpType
- Move RegexpType to the backend module.
- Add type alias for complex types.
- Remove clippy allows for type complexity.
